% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeseries_summary.R
\name{compute_monthly_deaths}
\alias{compute_monthly_deaths}
\title{compute_monthly_deaths}
\usage{
compute_monthly_deaths(.df, ..., .include_covid = FALSE)
}
\arguments{
\item{.df}{a tibble containing the variable date.of.death}

\item{...}{additional variables by which the monthly data should be grouped}

\item{.include_covid}{TRUE/FALSE (default TRUE) flag indicating whether to create a column specifically counting covid deaths}
}
\value{
a tibble containing the month of death, all grouping columns, and a count of deaths (and, optionally, an additional covid death count). Note that the month of death will be returned as a Date and all grouping variables will be returned as factors.
}
\description{
compute_monthly_deaths
}
\examples{
\dontrun{df_monthly_by_sex <- my_df \%>\%
                                            compute_monthly_deaths(.df = .,
                                                                         sex)}
}
