% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{create_covid_death_flag}
\alias{create_covid_death_flag}
\title{create_covid_death_flag}
\usage{
create_covid_death_flag(df, ucod_only = FALSE)
}
\arguments{
\item{df}{a tibble including the all record.axis.code variables and the final.cause.of.death.icd variable
from the raw death data}

\item{ucod_only}{Boolean determining whether to use only the underlying cause of death code ICD-10
or to use all available cause of death codes in the record.axis.code fields. TRUE means only
the underlying COD is used. Default is FALSE.}
}
\value{
tibble identical to input but with 1/0 covid_death flag
}
\description{
Uses all record.axis.code variables and the final.cause.of.death.icd variable to create a 1/0
flag where 1 means the death was due to covid. This code was taken from Matt Kiang's repo
here: https://github.com/mkiang/excess_drug_overdoses/blob/main/code/01_ingest_death_data.R
}
\examples{
\dontrun{df_covid_deaths_only = df \%>\%
                                        create_covid_death_flag() \%>\%
                                        filter(covid_death == 1)}
}
