% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_ca_region}
\alias{get_ca_region}
\title{get_ca_region
Joins a tibble that contains `county.of.death.code` as an integer with no leading zeroes
and joins to `ca_regions_fips` return the same tibble with additional variables:
 - `region` (number)
 - `region_name`
 - `county_name` (camel case)
 - `county_fips`
 - `county_cdph`}
\usage{
get_ca_region(df)
}
\arguments{
\item{df}{a tibble containing the variable `county.of.death.code` from the california death
data as an integer with no leading zeroes}
}
\value{
tibble same as input with additional region and county info
}
\description{
get_ca_region
Joins a tibble that contains `county.of.death.code` as an integer with no leading zeroes
and joins to `ca_regions_fips` return the same tibble with additional variables:
 - `region` (number)
 - `region_name`
 - `county_name` (camel case)
 - `county_fips`
 - `county_cdph`
}
\examples{
\dontrun{df_cleaned_county <- df \%>\%
                                          get_ca_region()}
}
