% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_ca_residence_only}
\alias{get_ca_residence_only}
\title{get_ca_only}
\usage{
get_ca_residence_only(df)
}
\arguments{
\item{df}{a cleaned version of the raw mortality file that includes the variable state_name}
}
\value{
tibble same as input df but with only rows where residence.state.province is california
}
\description{
get_ca_only
}
\examples{
\dontrun{
df_clean_ca <- df_clean \%>\%
                get_ca_only()}
}
