% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nioccs_ind_occ.R
\name{get_nioccs_ind_occ}
\alias{get_nioccs_ind_occ}
\title{get_nioccs_ind_occ}
\usage{
get_nioccs_ind_occ(df, number_results = 1, return_code_types = 2, version = 18)
}
\arguments{
\item{df}{a tibble containing columns called 'industry' and 'occupation' corresponding to the text values of industry and occupation}

\item{number_results}{Number of candidates returned. The default is 1, returning the top industry and occupation code based on the probabilities from the auto-coder machine learning models.}

\item{return_code_types}{Flag that determines what type of codes to return. The default is 2. Options are: c=0 returns NAICS 2017/SOC 2018 codes; c=1 returns Census Industry and Occupation 2018 codes; c=2 returns both sets of codes (NAICS/SOC and corresponding Census I/O)}

\item{version}{Determines the coding scheme version returned.Default is 18. Options are: v=18 returns codes from CDC Census 2018/CDC NAICS 2017/CDC SOC 2018 coding scheme; v=12 returns codes from the CDC Census 2012/CDC NAICS 2012/CDC SOC 2010 coding schemes}
}
\value{
tibble containing the original occupation and industry plus all results from NIOCCS API
}
\description{
get_nioccs_ind_occ
}
\examples{
\dontrun{
df_mortality_occupation <- tribble(~industry, ~occupation,
                                   "HOSPITAL", "NURSE",
                                   "EDUCATION", "MEDIA DIRECTOR")
df_nioccs <- get_nioccs_ind_occ(df = df_mortality_occupation,
                                  number_results = 3)
}
}
