% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_industry_occupation.R
\name{join_industry_occupation}
\alias{join_industry_occupation}
\title{join_industry_occupation}
\usage{
join_industry_occupation(df)
}
\arguments{
\item{df}{to match industry/occupation data to. Must contain variables `industry` and `occupation`.
See NIOCCS documentation for variable descriptions https://csams.cdc.gov/nioccs/HelpOutput.aspx

 - census_ind_code
 - census_ind_title
 - census_occ_code
 - census_occ_title
 - naics_code
 - naics_title
 - naics_probability
 - soc_code
 - soc_title
 - soc_probability
 - unexpected_naics_soc_combo}
}
\value{
tibble identical to input data plus 11 new variables:
}
\description{
This function takes a tibble from the mortality data (must include `industry` and `occupation` varaibles)
and joins NIOCCS-derived industry and occupation data.
}
\examples{
\dontrun{
my_df_with_ind_occ <- join_industry_occupation(my_df)}
}
