% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mortality_data.R
\name{load_ca_mortality_data}
\alias{load_ca_mortality_data}
\title{load_ca_mortality_data}
\usage{
load_ca_mortality_data(df = "latest")
}
\arguments{
\item{df}{a string specifying a specific mortality data set filename or latest (defaut)}
}
\value{
tibble
}
\description{
load_ca_mortality_data
}
\examples{
\dontrun{
  df_mortality_2022 <- load_ca_mortality_data(df = '2021-12-31')
  df_mortality_latest <- load_ca_mortality_data()
}
}
