% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_industry_occupation.R
\name{load_industry_occupation}
\alias{load_industry_occupation}
\title{load_industry_occupation}
\usage{
load_industry_occupation()
}
\value{
tibble containing all distinct industry + occupation combinations from the 2023-08-31 mortality file
and their matched census, naics, and soc codes/titles where possible through the NIOCCS matching online tool.
}
\description{
Convenience function for loading the industry and occupation code data associated with the 2023-08-31 mortality data file.
Matched codes from NIOCCS were pulled on March 28, 2025.
}
\examples{
\dontrun{df_ind_occ <- load_industry_occupation()}
}
