% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{min_data}
\alias{min_data}
\title{min_data
Remove unnecessary (potentially identifying) columns}
\usage{
min_data(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data}
}
\value{
tibble same as input with possibly identifying columns removed
}
\description{
min_data
Remove unnecessary (potentially identifying) columns
}
\examples{
\dontrun{
df_deidentified <- df_raw \%>\%
                      min_data()}
}
