% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nioccs_ind_occ.R
\name{nioccs_request}
\alias{nioccs_request}
\title{nioccs_request
This function makes an API request for 1 combination of industry + occupation}
\usage{
nioccs_request(i, o, n, c, v, u)
}
\arguments{
\item{i}{industry}

\item{o}{occupation}

\item{n}{number requested results per industy/occupation combination}

\item{c}{code type}

\item{v}{coding version}

\item{u}{unexpected result}
}
\value{
tibble
}
\description{
nioccs_request
This function makes an API request for 1 combination of industry + occupation
}
