% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{recode_age}
\alias{recode_age}
\title{recode_age}
\usage{
recode_age(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data}
}
\value{
tibble same as input with new variable age_int which is numeric age
}
\description{
recode_age
}
\examples{
\dontrun{
df_raw_with_int_age <- df_raw \%>\%
                          recode_age()}
}
