% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{recode_education}
\alias{recode_education}
\title{recode_education}
\usage{
recode_education(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data}
}
\value{
tibble same as input with new recoded education variable
}
\description{
recode_education
}
\examples{
\dontrun{
df_raw_with_clean_educ <- df_raw \%>\%
                            recode_education()}
}
