% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{recode_education_farmworkers}
\alias{recode_education_farmworkers}
\title{recode_education_farmworkers}
\usage{
recode_education_farmworkers(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data}
}
\value{
tibble same as input with new recoded education variable `recode_educ_farmwork`
}
\description{
recode_education_farmworkers
}
\examples{
\dontrun{
df_raw_with_clean_educ <- df_raw \%>\%
                            recode_education_farmworkers()}
}
