% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{recode_facility_of_death}
\alias{recode_facility_of_death}
\title{recode_facility_of_death}
\usage{
recode_facility_of_death(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data}
}
\value{
tibble same as input with new recode_facility_of_death variable
}
\description{
recode_facility_of_death
}
\examples{
\dontrun{
df_with_clean_race <- df_raw \%>\%
                          recode_facility_of_death()}
}
