% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{recode_marital_status}
\alias{recode_marital_status}
\title{recode_marital_status}
\usage{
recode_marital_status(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data with variable `marital.status`}
}
\value{
identical tibble with new variable `recode_marital_status`
}
\description{
recode_marital_status
}
\examples{
\dontrun{
df_with_clean_marital_status <- df_raw \%>\%
                                      recode_marital_status()}
}
