% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{recode_race}
\alias{recode_race}
\title{recode_race}
\usage{
recode_race(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data}
}
\value{
tibble same as input with new recoded race variable
}
\description{
recode_race
}
\examples{
\dontrun{
df_with_clean_race <- df_raw \%>\%
                          recode_race()}
}
