% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{recode_sex}
\alias{recode_sex}
\title{recode_sex}
\usage{
recode_sex(df)
}
\arguments{
\item{df}{a tibble containing raw mortality data}
}
\value{
tibble same as input with new recoded sex variable
}
\description{
recode_sex
}
\examples{
\dontrun{
df_raw_with_clean_sex <- df_raw \%>\%
                          recode_sex()}
}
