% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{subset_death_date}
\alias{subset_death_date}
\title{subset_death_date
Limits to records before or after (or both) certain dates (in yyyy-mm-dd format).
Note that for excess mortality and other algorithms, you might want to choose nearest sundays.}
\usage{
subset_death_date(df, before_date = "", after_date = "", remove_dod_na = F)
}
\arguments{
\item{df}{a tibble containing raw mortality data (including variable date.of.death)}

\item{before_date}{a string specifying a yyyy-mm-dddate. Only records with a date of death on or before this date will be kept.}

\item{after_date}{a string specifying a yyyy-mm-dd date. Only records with a date of death on or after this date will be kept.}

\item{remove_dod_na}{boolean where if False (default) does nothing. If True removes records where the date.of.death is NA}
}
\value{
tibble same columns as input with rows filtered on death dates provided
}
\description{
subset_death_date
Limits to records before or after (or both) certain dates (in yyyy-mm-dd format).
Note that for excess mortality and other algorithms, you might want to choose nearest sundays.
}
\examples{
\dontrun{
df_deidentified <- df_raw \%>\%
                      subset_death_date(before_date = "2023-08-31",
                                        after_date = "2019-06-21")}
}
